requirements.txt
# Core
requests
fastapi
uvicorn[standard]
python-multipart
websockets
python-dotenv
python-multipart
websockets

# AI/ML
#torch
transformers
accelerate
pillow
numpy
opencv-python-headless
faster-whisper

# Vector Database
pymilvus

# File handling
aiofiles

# Development
pydantic
